Challenge: convert Python dicitonary to XML
The key of the dictionary should go in the opening and closing tags, and
the value should go inside the tag. Note that if the value is actually a
`list` itself, each list item must go between opening and closing `item`
tags. Check the test file for a few examples.